automation/github_bot.py
"""
GitHub Bot for ContractorFinder Agent
Creates PR to MCP Registry
"""

import os
import json
from typing import Dict

try:
    from github import Github
    GITHUB_AVAILABLE = True
except ImportError:
    GITHUB_AVAILABLE = False
    print("Warning: PyGithub not installed. Using mock mode.")

class GitHubBot:
    """
    Submits ContractorFinder to MCP Registry
    """
    
    MCP_REGISTRY_REPO = "modelcontextprotocol/servers"
    
    def __init__(self):
        self.token = os.getenv("GITHUB_TOKEN")
        self.github = None
        
        if GITHUB_AVAILABLE and self.token:
            self.github = Github(self.token)
    
    def create_registry_pr(self) -> Dict:
        """
        Create PR to add ContractorFinder to MCP Registry
        """
        
        pr_content = {
            "title": "Add ContractorFinder - Licensed contractor search agent with x402 payments",
            "body": """## Description
ContractorFinder is an AI agent that finds licensed contractors with verified reviews. Built with x402 payments on Base blockchain.

## Features
- 🔍 Search licensed contractors by trade and location
- ✅ Verify license status (active/expired/suspended)
- ⭐ Pull real reviews and ratings
- 💰 $0.10 USDC per search via x402
- ⛓️ Built on Base blockchain
- 🛠️ MCP protocol compatible

## Tech Stack
- Python + FastAPI
- OpenAI GPT-4
- x402 payment protocol
- Apify for data scraping
- Deployed on Railway

## Live Endpoint
https://ca-contractor-finder-production.up.railway.app

## Pricing
$0.10 USDC per API call on Base mainnet

## Repository
https://github.com/yourname/contractor-finder-agent

## Checklist
- [x] Server is production-ready
- [x] Uses MCP protocol
- [x] Has clear documentation
- [x] Includes error handling
- [x] Has pricing model defined
""",
            "head_branch": "add-contractor-finder",
            "file_changes": {
                "src/contractor-finder/README.md": """# ContractorFinder

Find licensed contractors with verified reviews. Built with x402 payments on Base.

## Features
- Search contractors by trade (plumber, electrician, roofer, etc.)
- Verify license status
- Get reviews and ratings
- Pay $0.10 USDC per search

## Installation

```bash
pip install -r requirements.txt
```

## Usage

```python
from contractor_finder import ContractorFinderAgent

agent = ContractorFinderAgent()
result = agent.find_contractors("plumber", "Austin, TX")
```

## API

```bash
curl -X POST https://ca-contractor-finder-production.up.railway.app/find \\
  -H "Content-Type: application/json" \\
  -H "X-Payment-Signature: <x402_payment>" \\
  -d '{"trade": "plumber", "location": "Austin, TX"}'
```

## Pricing
$0.10 USDC per search on Base blockchain

## License
MIT
"""
            }
        }
        
        if not self.github:
            print("GitHub token not configured. Printing PR content:")
            print(f"\nTitle: {pr_content['title']}")
            print(f"\nBody:\n{pr_content['body']}")
            return {"status": "mock", "title": pr_content["title"]}
        
        try:
            # Get the MCP registry repo
            repo = self.github.get_repo(self.MCP_REGISTRY_REPO)
            
            # Create a new branch
            base_ref = repo.get_branch("main")
            repo.create_git_ref(
                ref=f"refs/heads/{pr_content['head_branch']}",
                sha=base_ref.commit.sha
            )
            
            # Create files (simplified - would need actual file creation)
            # This is a placeholder for the actual implementation
            
            # Create PR
            pr = repo.create_pull(
                title=pr_content["title"],
                body=pr_content["body"],
                head=pr_content["head_branch"],
                base="main"
            )
            
            return {
                "status": "success",
                "pr_number": pr.number,
                "pr_url": pr.html_url
            }
            
        except Exception as e:
            return {"status": "error", "error": str(e)}
    
    def create_repo_release(self, version: str = "v1.0.0") -> Dict:
        """Create a GitHub release"""
        
        if not self.github:
            print(f"Would create release {version}")
            return {"status": "mock", "version": version}
        
        try:
            # Get the user's repo
            user = self.github.get_user()
            repo = user.get_repo("contractor-finder-agent")
            
            release = repo.create_git_release(
                tag=version,
                name=f"ContractorFinder {version}",
                message="Initial release - AI agent for finding licensed contractors with x402 payments",
                draft=False,
                prerelease=False
            )
            
            return {
                "status": "success",
                "release_url": release.html_url,
                "tag": version
            }
            
        except Exception as e:
            return {"status": "error", "error": str(e)}

def main():
    """Run the GitHub bot"""
    bot = GitHubBot()
    
    print("🐙 GitHub Bot for ContractorFinder")
    print("=" * 50)
    
    # Create registry PR
    print("\n📋 Creating MCP Registry PR...")
    result = bot.create_registry_pr()
    
    if result["status"] == "success":
        print(f"✅ PR created: {result['pr_url']}")
    elif result["status"] == "mock":
        print("✅ Mock mode: PR content ready")
    else:
        print(f"❌ Error: {result.get('error')}")
    
    # Create release
    print("\n🏷️ Creating GitHub release...")
    result = bot.create_repo_release()
    
    if result["status"] == "success":
        print(f"✅ Release created: {result['release_url']}")
    elif result["status"] == "mock":
        print("✅ Mock mode: Release ready")
        print("\nTo create for real, set GitHub token in .env:")
        print("  GITHUB_TOKEN")

if __name__ == "__main__":
    main()
