requirements.txt
openai>=1.0.0
mcp>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
tweepy>=4.14.0
praw>=7.7.0
PyGithub>=2.1.0
